Remove `UiLayer.OnFocused()` and replace with `GrabFocus()` virtual
