+++ OpenNefia.Content/CharaMake/CharaMakeAppearanceLayer.cs
@@ -70,9 +70,9 @@ namespace OpenNefia.Content.CharaMake
             }
         }
 
-        public override void OnFocused()
+        public override void GrabFocus()
         {
-            base.OnFocused();
+            base.GrabFocus();
             List.GrabFocus();
         }
