+++ OpenNefia.Content/Equipment/EquipmentLayer.cs
@@ -236,9 +236,9 @@ namespace OpenNefia.Content.Equipment
             UpdateFromEquipTarget();
         }
 
-        public override void OnFocused()
+        public override void GrabFocus()
         {
-            base.OnFocused();
+            base.GrabFocus();
             List.GrabFocus();
         }
