+++ OpenNefia.Content/TitleScreen/TitleScreenLayer.cs
@@ -154,9 +154,9 @@ namespace OpenNefia.Content.TitleScreen
             return keyHints;
         }
 
-        public override void OnFocused()
+        public override void GrabFocus()
         {
-            base.OnFocused();
+            base.GrabFocus();
             List.GrabFocus();
         }
